x64 bug fixed

+++ SSEngine/Engine/Controller/Private/ControllerManager.cpp
@@ -5,8 +5,8 @@ ControllerManager *ControllerManager::globalInstance_ = nullptr;
 
 void ControllerManager::RemoveMonsterController(IController &i_monsterController)
 {
-	int32 size = monsterControllers_.size();
-	for (int32 i = 0; i < size; ++i)
+	uint32 size = static_cast<uint32>(monsterControllers_.size());
+	for (uint32 i = 0; i < size; ++i)
 	{
 		if (monsterControllers_[i])
 		{
@@ -22,8 +22,8 @@ void ControllerManager::RemoveMonsterController(IController &i_monsterController
 
 void ControllerManager::UpdatePlayerController()
 {
-	int size = playerControllers_.size();
-	for (int i = 0;i < size;++i)
+	uint32 size = static_cast<uint32>(playerControllers_.size());
+	for (uint32 i = 0;i < size;++i)
 	{
 		playerControllers_[i]->UpdateController();
 	}
@@ -31,8 +31,8 @@ void ControllerManager::UpdatePlayerController()
 
 void ControllerManager::UpdateMonsterController()
 {
-	int size = monsterControllers_.size();
-	for (int i = 0;i < size;++i)
+	int32 size = static_cast<uint32>(monsterControllers_.size());
+	for (int32 i = 0;i < size;++i)
 	{
 		if ((*monsterControllers_[i]).IsValid())
 		{
